# api/openapi.yaml
openapi: 3.0.3
info:
  title: Memory Gateway API (RAG-only)
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: ok
  /memories/search:
    post:
      summary: Search memories
      description: Returns hybrid-ranked memories matching the provided query.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySearchRequest'
      responses:
        '200':
          description: Successful search response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySearchResponse'
  /memories/retrieve_for_task:
    post:
      summary: Retrieve memories for task
      description: Returns contextual memories to support the provided task description.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrieveForTaskRequest'
      responses:
        '200':
          description: Successful task retrieval response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveForTaskResponse'
components:
  schemas:
    MemoryPayload:
      type: object
      required:
        - memory_id
        - title
        - branch
        - content
        - salience
        - memory_type
      properties:
        memory_id:
          type: string
          description: Stable identifier for the memory record.
        title:
          type: string
          description: Short label summarising the memory.
        branch:
          type: string
          description: Logical branch or folder the memory belongs to.
        content:
          type: string
          description: Full text content associated with the memory.
        salience:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Relative salience score (0-1).
        memory_type:
          type: string
          enum:
            - semantic
            - episodic
            - procedural
          description: Categorisation of the memory (semantic, episodic, procedural).
    MemorySearchHit:
      type: object
      required:
        - memory
        - score
      properties:
        memory:
          $ref: '#/components/schemas/MemoryPayload'
        score:
          type: number
          format: float
          minimum: 0
          description: Hybrid keyword + salience score.
    MemorySearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          description: Free-text query to match memories.
        limit:
          type: integer
          default: 5
          minimum: 1
          maximum: 20
          description: Maximum number of memories to return (default 5, max 20).
    MemorySearchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/MemorySearchHit'
    RetrieveForTaskRequest:
      type: object
      required:
        - task
      properties:
        task:
          type: string
          minLength: 1
          description: Task description requiring contextual memories.
        branch:
          type: string
          nullable: true
          description: Branch hint to narrow retrieval to specific workstreams.
        limit:
          type: integer
          default: 3
          minimum: 1
          maximum: 20
          description: Maximum number of supporting memories to return (default 3, max 20).
    RetrieveForTaskResponse:
      type: object
      required:
        - task
        - context
        - provenance
      properties:
        task:
          type: string
        context:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/MemoryTypeGroup'
        provenance:
          type: array
          items:
            $ref: '#/components/schemas/Provenance'
    Provenance:
      type: object
      required:
        - memory_id
        - memory_type
        - detail
      properties:
        memory_id:
          type: string
          description: Identifier for the memory the provenance relates to.
        memory_type:
          type: string
          enum:
            - semantic
            - episodic
            - procedural
          description: Memory type tied to the provenance record.
        detail:
          type: string
          description: Source or lineage description for the memory.
    MemoryTypeGroup:
      type: object
      required:
        - hits
        - provenance
      properties:
        hits:
          type: array
          items:
            $ref: '#/components/schemas/MemorySearchHit'
          description: Ranked memories for the type.
        provenance:
          type: array
          items:
            $ref: '#/components/schemas/Provenance'
          description: Provenance metadata for memories within the type group.